backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1
pydantic==2.3.0
websockets==11.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.20
# Add your additional dependencies here
